refactor(db): renamed `schema_2.ts` to `schema.ts` and removed unneccessary lines from `drizzle.ts`
